Move make_folders back into convert.py * Minimize CSS size by installing CSS into css/categories * A few hacks to integrate with the HTML
