+++ index.html
@@ -1,5 +1,6 @@
 <head>
   <link rel="stylesheet" href="css/nicktardif.min.css">
+  <link rel="stylesheet" href="css/categories/Recent.css">
   <script src="js/nicktardif.min.js"></script>
   <script src="js/index.js"></script>
   <script>
